Move relevant ads endpoint to weapons plugin
