+++ imagespace/server/imagesearch_rest.py
@@ -21,9 +21,7 @@ from girder import events
 from girder.api import access
 from girder.api.describe import Description
 from girder.api.rest import Resource
-from urlparse import urlparse
 
-import itertools
 import json
 import requests
 import os
@@ -33,56 +31,6 @@ class ImageSearch(Resource):
     def __init__(self):
         self.resourceName = 'imagesearch'
         self.route('GET', (), self.getImageSearch)
-        self.route('GET', ('relevant_ads',), self.getRelevantAds)
-
-    @access.public
-    def getRelevantAds(self, params):
-        AD_LIMIT = 20
-        def sort_documents_by_url(documents):
-            return sorted(documents, key=lambda x: x['url'])
-
-        # @todo this assumes all URLs returned from Solr will properly urlparse
-        def group_documents_by_domain(documents):
-            return itertools.groupby(sort_documents_by_url(documents),
-                                     lambda doc: urlparse(doc['url']).netloc)
-
-        try:
-            result = requests.get(os.environ['IMAGE_SPACE_SOLR'] + '/select', params={
-                'wt': 'json',
-                'q': 'outpaths:"%s"' % params['solr_image_id'],
-                'fl': 'id,url',
-                'rows': str(AD_LIMIT)
-            }, verify=False).json()
-        except ValueError:
-            return {
-                'numFound': 0,
-                'docs': []
-            }
-
-        try:
-            response = {
-                'numFound': result['response']['numFound'],
-                'docs': result['response']['docs'],
-                'groupedDocs': []
-            }
-        except KeyError:
-            return {
-                'numFound': 0,
-                'docs': []
-            }
-
-        for (domain, documents) in group_documents_by_domain(response['docs']):
-            response['groupedDocs'].append([domain, list(documents)])
-
-        # Display the domain with the largest number of documents first
-        response['groupedDocs'] = sorted(response['groupedDocs'],
-                                         key=lambda (_, docs): len(docs),
-                                         reverse=True)
-
-        return response
-    getRelevantAds.description = Description(
-        'Retrieve the relevant ad ids and urls from a given image'
-    ).param('solr_image_id', 'ID of the Solr document representing an image')
 
     @access.public
     def getImageSearch(self, params):
